# internal/models/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package models

import (
	"time"

	uuid "github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Category struct {
	ID        int32     `json:"id"`
	Name      string    `json:"name"`
	CreatedAt time.Time `json:"created_at"`
}

type Event struct {
	ID           uuid.UUID   `json:"id"`
	Title        string      `json:"title"`
	Description  string      `json:"description"`
	Status       string      `json:"status"`
	CategoryID   int32       `json:"category_id"`
	OrganizerID  uuid.UUID   `json:"organizer_id"`
	Location     pgtype.Text `json:"location"`
	StartingTime time.Time   `json:"starting_time"`
	EndingTime   time.Time   `json:"ending_time"`
	CreatedAt    time.Time   `json:"created_at"`
	UpdatedAt    time.Time   `json:"updated_at"`
}

type User struct {
	ID        uuid.UUID `json:"id"`
	Auth0ID   string    `json:"auth0_id"`
	CreatedAt time.Time `json:"created_at"`
}